project update based on comments
