+++ backend/Dockerfile
@@ -1,5 +1,5 @@
 FROM openjdk:11-jdk-slim
-
+MAINTAINER Nikolay Petryashev <petryashev.n.o@gmail.com>
 ADD target/backend-hw.jar backend.jar
 ADD src/etc/service.properties etc/service.properties
 ADD src/etc/hibernate.properties etc/hibernate.properties
