+++ backend/init.sql
@@ -1,54 +1,54 @@
-DROP TABLE resume;
-DROP TABLE vacancy;
-DROP TABLE company;
-DROP TABLE userHH;
+-- DROP TABLE resume;
+-- DROP TABLE vacancy;
+-- DROP TABLE company;
+-- DROP TABLE users;
+CREATE TYPE user_type AS ENUM ('employer', 'applicant');
 
-CREATE TABLE userHH(
-     user_id SERIAL PRIMARY KEY,
+CREATE TABLE users(
+     users_id SERIAL PRIMARY KEY,
      name VARCHAR(256),
-     user_type integer DEFAULT 0 NOT NULL,
-     creation_date timestamp NOT NULL,
-     update_date timestamp NOT NULL,
-     CONSTRAINT user_type_validate CHECK ((user_type = ANY (ARRAY[0, 1])))
+     user_type user_type NOT NULL,
+     creation_date TIMESTAMP NOT NULL,
+     update_date TIMESTAMP NOT NULL
 );
 
 CREATE TABLE resume(
      resume_id SERIAL PRIMARY KEY,
-     user_id INTEGER DEFAULT 0 NOT NULL,
-     title varchar(256) DEFAULT ''::varchar NOT NULL,
-     work_experience text,
-     contacts text,
-     creation_date timestamp NOT NULL,
-     update_date timestamp NOT NULL,
-     FOREIGN KEY (user_id) REFERENCES userHH(user_id)
+     users_id INTEGER NOT NULL,
+     title VARCHAR(256) DEFAULT ''::VARCHAR NOT NULL,
+     work_experience TEXT,
+     contacts TEXT,
+     creation_date TIMESTAMP NOT NULL,
+     update_date TIMESTAMP NOT NULL,
+     FOREIGN KEY (users_id) REFERENCES users(users_id)
 );
 
 CREATE TABLE company(
      company_id SERIAL PRIMARY KEY,
-     user_id INTEGER DEFAULT 0 NOT NULL,
-     title varchar(256) DEFAULT ''::varchar NOT NULL,
-     creation_date timestamp NOT NULL,
-     update_date timestamp NOT NULL,
-     FOREIGN KEY (user_id) REFERENCES userHH(user_id)
+     users_id INTEGER NOT NULL,
+     title VARCHAR(256) DEFAULT ''::VARCHAR NOT NULL,
+     creation_date TIMESTAMP NOT NULL,
+     update_date TIMESTAMP NOT NULL,
+     FOREIGN KEY (users_id) REFERENCES users(users_id)
 );
 
 CREATE TABLE vacancy(
      vacancy_id SERIAL PRIMARY KEY,
-     company_id INTEGER DEFAULT 0 NOT NULL,
-     title varchar(256) DEFAULT ''::varchar NOT NULL,
+     company_id INTEGER NOT NULL,
+     title VARCHAR(256) DEFAULT ''::VARCHAR NOT NULL,
      compensation bigint DEFAULT 0,
-     description text,
-     contacts text,
-     creation_date timestamp NOT NULL,
-     update_date timestamp NOT NULL,
+     description TEXT,
+     contacts TEXT,
+     creation_date TIMESTAMP NOT NULL,
+     update_date TIMESTAMP NOT NULL,
      FOREIGN KEY (company_id) REFERENCES company(company_id)
 );
 
-INSERT INTO userHH (name, user_type, creation_date, update_date) VALUES
-    ('Павел Иванов', 0, now(), now()),
-    ('Татьяна Семеновна', 0, now(), now()),
-    ('Сергей Сидоров', 0, now(), now()),
-    ('Александр Петров', 1, now(), now());
+INSERT INTO users (name, user_type, creation_date, update_date) VALUES
+    ('Павел Иванов', 'applicant', now(), now()),
+    ('Татьяна Семеновна', 'applicant', now(), now()),
+    ('Сергей Сидоров', 'applicant', now(), now()),
+    ('Александр Петров', 'employer', now(), now());
 
 INSERT INTO resume (title, user_id, work_experience, contacts, creation_date, update_date) VALUES
     ('Программист Java', 1, '3 года', 'hh1@mail.com', now(), now()),
