+++ backend/src/main/java/ru/hh/school/dao/CompanyDao.java
@@ -2,32 +2,30 @@ package ru.hh.school.dao;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-import ru.hh.school.models.Company;
+import ru.hh.school.entity.Company;
 
+import javax.inject.Inject;
+import javax.inject.Named;
 import java.util.List;
 
-@Repository
+@Named
 public class CompanyDao {
-    private final SessionFactory sessionFactory;
 
+    private final SessionFactory sessionFactory;
+    @Inject
     public CompanyDao(SessionFactory sessionFactory) {
         this.sessionFactory = sessionFactory;
     }
 
-    @Transactional
     public Company create(Company company) {
         session().save(company);
         return company;
     }
 
-    @Transactional
     public Company get(Integer id) {
         return session().get(Company.class, id);
     }
 
-    @Transactional
     public List<Company> getAll() {
         return session().createQuery("FROM Company", Company.class)
                 .getResultList();
