+++ backend/src/main/java/ru/hh/school/dao/ResumeDao.java
@@ -1,40 +1,37 @@
 package ru.hh.school.dao;
 
-import org.springframework.stereotype.Repository;
-import ru.hh.school.models.Resume;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.springframework.transaction.annotation.Transactional;
+import ru.hh.school.entity.Resume;
 
+import javax.inject.Inject;
+import javax.inject.Named;
 import java.util.List;
 
-@Repository
+@Named
 public class ResumeDao {
-    private final SessionFactory sessionFactory;
 
+    private final SessionFactory sessionFactory;
+    @Inject
     public ResumeDao(SessionFactory sessionFactory) {
         this.sessionFactory = sessionFactory;
     }
 
-    @Transactional
     public Resume create(Resume resume) {
         session().persist(resume);
         return resume;
     }
 
-    @Transactional
     public Resume get(Integer id) {
         return session().get(Resume.class, id);
     }
 
-    @Transactional
     public List<Resume> getByUserId(Integer userId) {
         return session().createQuery("FROM Resume WHERE userId=:userId", Resume.class)
                         .setParameter("userId", userId)
                         .getResultList();
     }
 
-    @Transactional
     public List<Resume> getAll() {
         return session().createQuery("FROM Resume", Resume.class)
                 .getResultList();
