+++ backend/src/main/java/ru/hh/school/dao/UserDao.java
@@ -1,58 +1,52 @@
 package ru.hh.school.dao;
 
-import org.springframework.stereotype.Repository;
-import ru.hh.school.models.User;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.springframework.transaction.annotation.Transactional;
+import ru.hh.school.entity.User;
+import ru.hh.school.entity.UserType;
 
+import javax.inject.Inject;
+import javax.inject.Named;
 import java.util.List;
 
-@Repository
+@Named
 public class UserDao {
 
     private final SessionFactory sessionFactory;
-
+    @Inject
     public UserDao(SessionFactory sessionFactory) {
         this.sessionFactory = sessionFactory;
     }
 
-    @Transactional
     public User create(User user) {
         session().save(user);
         return user;
     }
 
-    @Transactional
     public List<User> getAll() {
         return session().createQuery("FROM User", User.class)
                 .getResultList();
     }
 
-    @Transactional
     public User get(Integer id) {
         return session().get(User.class, id);
     }
 
-    @Transactional
-    public List<User> getByType(Integer userType) {
+    public List<User> getByType(UserType userType) {
         return session().createQuery("FROM User WHERE userType = :userType", User.class)
                 .setParameter("userType", userType)
                 .getResultList();
     }
 
-    @Transactional
     public void deleteBy(Integer id) {
       session().createQuery("DELETE FROM User WHERE id = :id_to_delete")
               .setParameter("id_to_delete", id).executeUpdate();
     }
 
-    @Transactional
     public void deleteAll() {
         session().createQuery("DELETE FROM User").executeUpdate();
     }
 
-    @Transactional
     public void update(User user) {
         session().update(user);
     }
