+++ backend/src/main/java/ru/hh/school/dao/VacancyDao.java
@@ -2,39 +2,36 @@ package ru.hh.school.dao;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.springframework.stereotype.Repository;
-import ru.hh.school.models.Vacancy;
-import org.springframework.transaction.annotation.Transactional;
+import ru.hh.school.entity.Vacancy;
 
+import javax.inject.Inject;
+import javax.inject.Named;
 import java.util.List;
 
-@Repository
+@Named
 public class VacancyDao {
-    private final SessionFactory sessionFactory;
 
+    private final SessionFactory sessionFactory;
+    @Inject
     public VacancyDao(SessionFactory sessionFactory) {
         this.sessionFactory = sessionFactory;
     }
 
-    @Transactional
     public Vacancy create(Vacancy vacancy) {
         session().save(vacancy);
         return vacancy;
     }
 
-    @Transactional
     public Vacancy get(Integer id) {
         return session().get(Vacancy.class, id);
     }
 
-    @Transactional
     public List<Vacancy> getByCompanyId(Integer companyId) {
         return session().createQuery("FROM Vacancy WHERE companyId=:companyId", Vacancy.class)
                 .setParameter("companyId", companyId)
                 .getResultList();
     }
 
-    @Transactional
     public List<Vacancy> getAll() {
         return session().createQuery("FROM Vacancy", Vacancy.class)
                 .getResultList();
