+++ backend/src/main/java/ru/hh/school/entity/Company.java
@@ -1,31 +1,29 @@
-package ru.hh.school.models;
+package ru.hh.school.entity;
 
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
 import java.sql.Timestamp;
+import java.util.List;
 
 @Entity
-@Table(name = "resume")
-public class Resume {
+@Table(name = "company")
+public class Company {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column (name = "resume_id",  updatable = false)
+    @Column (name = "company_id",  updatable = false)
     private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn (name = "user_id")
+    @JoinColumn (name = "users_id")
     private User user;
+
+    @OneToMany(mappedBy = "company", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Vacancy> vacancies;
 
     @Column (name = "title")
     private String title;
-
-    @Column (name = "work_experience")
-    private String workExperience;
-
-    @Column (name = "contacts")
-    private String contacts;
 
     @CreationTimestamp
     @Column (name = "creation_date")
@@ -36,7 +34,7 @@ public class Resume {
     private Timestamp updateDate;
 
     // ToDo: no-arg constructor
-    public Resume (){
+    public Company (){
 
     }
 
@@ -56,28 +54,20 @@ public class Resume {
         this.user = user;
     }
 
+    public List<Vacancy> getVacancies() {
+        return vacancies;
+    }
+
+    public void setVacancies(List<Vacancy> vacancies) {
+        this.vacancies = vacancies;
+    }
+
     public String getTitle() {
         return title;
     }
 
     public void setTitle(String title) {
         this.title = title;
-    }
-
-    public String getWorkExperience() {
-        return workExperience;
-    }
-
-    public void setWorkExperience(String workExperience) {
-        this.workExperience = workExperience;
-    }
-
-    public String getContacts() {
-        return contacts;
-    }
-
-    public void setContacts(String contacts) {
-        this.contacts = contacts;
     }
 
     public Timestamp getCreationDate() {
