+++ backend/src/main/java/ru/hh/school/entity/UserType.java
@@ -1,0 +1,6 @@
+package ru.hh.school.entity;
+
+public enum UserType {
+    EMPLOYER,
+    APPLICANT
+}
