+++ backend/src/main/java/ru/hh/school/service/ResumeService.java
@@ -1,19 +1,18 @@
-package ru.hh.school.services;
+package ru.hh.school.service;
 
-import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.hh.school.dao.ResumeDao;
 import ru.hh.school.dao.UserDao;
 import ru.hh.school.dto.ResumeRequestDto;
-import ru.hh.school.models.Resume;
-import ru.hh.school.models.User;
+import ru.hh.school.entity.Resume;
+import ru.hh.school.entity.User;
+import ru.hh.school.entity.UserType;
 
 import javax.inject.Inject;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.Response;
-import java.time.Instant;
+import javax.inject.Singleton;
 import java.util.List;
 
-@Service
+@Singleton
 public class ResumeService {
     private final ResumeDao resumeDao;
     private final UserDao userDao;
@@ -24,9 +23,10 @@ public class ResumeService {
         this.userDao = userDao;
     }
 
+    @Transactional
     public void saveNew(ResumeRequestDto resumeDto) {
         User user = userDao.get(resumeDto.getUserId());
-        if (user.getUserType() == 0) {
+        if (user.getUserType() == UserType.APPLICANT) {
             Resume resume = new Resume();
             resume.setUser(user);
             resume.setTitle(resumeDto.getTitle());
@@ -34,19 +34,23 @@ public class ResumeService {
             resume.setContacts(resumeDto.getContacts());
             //TODO проверить, чтобы добавлялось время
             resumeDao.create(resume);
-        } else {
-            throw new WebApplicationException(Response.Status.NOT_FOUND);
         }
+//        else {
+//            throw new WebApplicationException(Response.Status.NOT_FOUND);
+//        }
     }
 
+    @Transactional
     public Resume getBy(int resumeId) {
         return resumeDao.get(resumeId);
     }
 
+    @Transactional
     public List<Resume> getAll() {
         return resumeDao.getAll();
     }
 
+    @Transactional
     public List<Resume> getResumesForUserId(int userId) {
         return resumeDao.getByUserId(userId);
     }
