+++ backend/src/main/java/ru/hh/school/service/UserService.java
@@ -1,14 +1,16 @@
-package ru.hh.school.services;
+package ru.hh.school.service;
 
-import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.hh.school.dao.UserDao;
 import ru.hh.school.dto.UserRequestDto;
-import ru.hh.school.models.User;
+import ru.hh.school.entity.User;
+import ru.hh.school.entity.UserType;
 
 import javax.inject.Inject;
+import javax.inject.Singleton;
 import java.util.List;
 
-@Service
+@Singleton
 public class UserService {
 
   private final UserDao userDao;
@@ -17,7 +19,7 @@ public class UserService {
   public UserService(UserDao userDao) {
     this.userDao = userDao;
   }
-
+  @Transactional
   public void saveNew(UserRequestDto userDto) {
     User user = new User();
     user.setName(userDto.getName());
@@ -25,23 +27,23 @@ public class UserService {
     //TODO проверить, чтобы добавлялось время
     userDao.create(user);
   }
-
-  public List<User> getUsersByType(Integer userType) {
+  @Transactional
+  public List<User> getUsersByType(UserType userType) {
     return userDao.getByType(userType);
   }
-
+  @Transactional
   public List<User> getAll() {
     return userDao.getAll();
   }
-
+  @Transactional
   public void deleteAll() {
     userDao.deleteAll();
   }
-
+  @Transactional
   public User getBy(int userId) {
     return userDao.get(userId);
   }
-
+  @Transactional
   public void deleteBy(int userId) {
     userDao.deleteBy(userId);
   }
