+++ frontend/nginx.conf
@@ -9,8 +9,8 @@ http {
     listen   80;
     root /usr/share/nginx/html;
 	
-	location /api/ {
-		proxy_pass http://hh-homework-backend:8080/api/;	
-	}
+    location /api/ {
+      proxy_pass http://hh-homework-backend:8080/api/;	
+    }
   }
 }
